+++ InventoryMgmt-Services/src/main/java/com/innovatehub/inventorymgmt/services/stock/ProductCategoryService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import com.innovatehub.inventorymgmt.common.model.stock.ProductCategory;
 
 public interface ProductCategoryService {
-	public int saveProductCategory(ProductCategory productCategory);
-	public ProductCategory getProductCategory(int productCategory);
+	public Long saveProductCategory(ProductCategory productCategory);
+	public ProductCategory getProductCategory(Long productCategory);
 	public List<ProductCategory> getAllProductCategories();
 }
